+++ frontend/app/common/components/directory-record-listitem.tsx
@@ -1,5 +1,5 @@
 //import type { Items } from "../fs/types";
-import type { Item } from "../newer-fs/types";
+import type { Item } from "../fs/types";
 
 export default function DirectoryRecordListItem({rec, children}: {rec: Item, children?: React.ReactNode}) {
     return (
