+++ frontend/app/host/server-com/message/writers.ts
@@ -1,7 +1,7 @@
 import { FlagService, encodePerJson, encodeUUID } from "../../../common/server-com/binary";
 import { type EncryptionData, encryptBuffer } from "../../../common/crypto";
 import { type HomeNodeFrontendConfig } from "../../../config";
-import type { Item } from "~/common/newer-fs/types";
+import type { Item } from "~/common/fs/types";
 
 
 export namespace HostToServerMessage {
